+++ src/index.ts
@@ -1,2 +1,6 @@
+import './theme/themeAugmentation';
+
 export * from './theme';
 export * from './helpers/commonTypes';
+
+export type { Theme } from '@mui/material';